🏅🌌 ↝ Clean up, create basic square generator based on parameters that match the planetsss table on postgres
