+++ Generator/generateCall.py
@@ -1,0 +1,7 @@
+# Call generator script
+# Create image locally
+# Send image as request
+# Once request is received, delete the image locally
+# Discuss scaling issues
+# Lightkurve .fits as well
+# And then, of course, look at ingress & egress points for desci-md/automating-metadata
