+++ Pipfile
@@ -1,20 +1,10 @@
 [packages]
-moralis = "*"
 flask = "*"ac
 flask_cors = "*"
-thirdweb-sdk = "*"
 python-dotenv = "*"
-psycopg2 = "*"
 matplotlib = "*"
 lightkurve = "*"
 nbformat = "*"
 gunicorn = "*"
-sqlalchemy = "*"
-psycopg2-binary = "*"
-flask-sqlalchemy = "*"
 supabase = "*"
-gotrue = "*"
-web3 = "*"
-brownie = "*"
-eth-brownie = "*"
-eth_utils = "*"
+noise = "*"
