+++ Pipfile.extra
@@ -1,0 +1,11 @@
+web3 = "*"
+brownie = "*"
+eth-brownie = "*"
+eth_utils = "*"
+thirdweb-sdk = "*"
+moralis = "*"
+psycopg2 = "*"
+psycopg2-binary = "*"
+flask-sqlalchemy = "*"
+sqlalchemy = "*"
+gotrue = "*"
