--- constants/contractMappingResponse.ts
@@ -1,8 +1,0 @@
-import { BigNumber } from "ethers";
-
-type ContractMappingResponse = {
-    isData: boolean;
-    value: BigNumber;
-};
-
-export default ContractMappingResponse;
