+++ context/constants/contractMappingResponse.ts
@@ -1,0 +1,8 @@
+import { BigNumber } from "ethers";
+
+type ContractMappingResponse = {
+    isData: boolean;
+    value: BigNumber;
+};
+
+export default ContractMappingResponse;
